Adjust some monster tracer behavior to be friendlier to the player
Previously, while monster tracers factored in the resists of other
monsters when calculating friend/foe hit values, they would never do this
for players. This meant that allied monsters wouldn't realize that bolts of
draining/venom were harmless to undead players and refuse to ever fire them
through the player (although this is a behavior we constantly see them do
with other monsters). Since it should be entirely helpful to let allies do
this, allied monsters now consult harmless_to_player() when determining
whether to worry about hitting you.

Additionally, allied monsters with penetrating beams of any kind that are
NOT harmless to you, will stop their shot immediately before it would hit
you - passing through as many monsters as possible up to that point. This
sort of behavior was attempted in 72347fce6 and then reverted in
98e7288b7fd because it stopped all allied piercing beams at their first
target regardless of whether this mattered. This commit attempts to be a
little more specific and only stop them when it DOES matter.

(Short-circuiting beam tracing like this feels just slightly sketchy to me,
but in preliminary testing, it appears to work fine)