Greatly increase Laughing Skull spawn rates in late-D and Crypt
For some reason, they were essentially the least common monsters in the
normal spawn list for both places (and vanishingly unlikely in D). If we're
going to add a shiny new monster, let's let people actually encounter them!

(Band placement means that the Crypt ones will be a lot nastier than D, if
anyone is curious from the numbers)