Have AF_ANTIMAGIC apply 50% more stacks to monsters with none
The antimagic formula is kind of lopsided, in that short durations of the
effect also have dramatically reduced effectiveness even for the brief
while it lasts. It's been suggested that the formula itself could use some
revisions, but I'm reluctant to touch something that upset the balance of
player antimagic weapons so close to freeze.

This commit is MOSTLY aimed at boosting Summon Mana Viper for players - it
doesn't affect the player behind hit by them, or antimagic weapons at all.
But ideally it makes it a bit easier for the viper to apply enough
antimagic to actually do something without also boosting its damage.

(Mana vipers themselves are slated for a possible rework in 0.32, but this
can act as a stopgap until then)