Very slightly trim Ice Cave and Ossuary ranges
Prior to the introduction of Depths, the Ice Cave difficulty split in D was
a coinflip on a random midpoint floor of D (D:14 or 15) and always set to
hard when deeper than that. The coinflip was removed in b4d244a, but
183a740 reintroduced the notion by making D:14-15 contain hard ice caves.
This was done due to the bottom of D being not very far away from having
to explore the Lair branches themselves, where those are always set to
hard. Over time, there's been confusion about which of those floors
contains hard ice caves- the D:13-14 threshold neither uses the very last
floor of D, nor does it stop at the thresholds use for placing Orc or
Vaults. It especially doesn't help that at the current juncture, in-game
nothing is actually explained about harder ice caves outside of noticing
what particular monsters one faces- it mostly matches to specific branch
ranges, but falters on this exact point.

While there's been discussion about using different names for features
and the sidebar for distinguishing this difficulty split ("ice grottos"?),
there's also been some discussion about how ice caves strangely stretch
out one element for a long chunk of the game itself, to the point of using
specific ugly thing colours, draconians, and monsters mostly unique to V
or Cocytus. It doesn't help that there's not any time gap between most of
when one can encounter easier versus hard ice caves, either. While I
actually have some interest in stretching out the easy versus hard portal
concept further in the future, most of it would feature much wider divides
than D:13 versus D:14, anyway. With all of this considered, I'm cutting out
the D:14 and D:15 ice cave spawns entirely- it should introduce a tiny bit
of a gap, reduce the pressure ice caves have for spawning so often, and
removing the confusion for 0.31 which floors of D can contain which ice
caves difficulties.

Also. Ossuaries are in a very awkward part of the game in some ways- a D:4
character and a D:8 character can be very far apart just in terms of
finding chain / plate, a non-starter weapon, or being able to reliably cast
a L3 or L4 spell instead of L2. While it's good to feature them early as a
portal to shake up the most-often-seen early game, and since they got so
such a recent rebalancing and set updates, later ossuaries still rarely
feature any much for threats. In lieu of adding idiosyncratic difficulty
scaling to the lot of them, I'm instead trimming the very last floor
of ossuaries for now- it makes their range the same amount of floors as
Sewers, and the deathrate for D:8 ossuaries over the past three versions
is a paltry ~0.5% anyway. This comes with very slightly more chance to
see Ossuaries on D:6 and D:7, to somewhat make up for the lowered overall
Ossuary count.