Trim a bit off of Aizul's deeper depths (Ge0FF)
Following the spirit of 005a376, Aizul's got good dangerous appearances in
the Lair branches and a pre / post Lair branches Elf, but falls off later
on as player XLs go up and innate Willpower can stave off her Sleep casts.
With around a ~0.3% or worse killratio in Vaults, Depths, and Crypt, and
with plenty of other late uniques around to provide variety, she can
slither away from all but the first two floors of Vaults and be better off.