Fix a terrible interaction between Noxious Bog and quicksilver
Casting Eringya's Noxious Bog simultaneously created bog tiles around the
player and gave a status that made the player immune to these, *but those
two things were not otherwise linked*. This meant that if the player lost
the status (due to drinking !cancellion or being hit by quicksilver), the
bog would remain, but they would lose their immunity to it - often swiftly
dying in the process.

Since the bog is supposed to be radiating from the player's magical effect
in the first place, instead immediately remove the bog if the player loses
the status prematurely.