Fix blastmote knockback (and maybe other things??) not clearing water hold
If the player was knocked back by a blastmote explosion, it would not end
a water elemental engulf that was currently in progress, even if it knocked
you away from them. Curiously, this wasn't true if the water elemental was
the one knocked back instead.

I am going to go on record here and say there are FAR TOO MANY similarly
named functions for moving players and monsters around, which check some
different subset of different properties (and sometimes call each other,
but sometimes don't). It's not at all clear to me which you are even
SUPPOSED to call in different situations (is it clear to anyone?) and given
the breadth and complexity of different checks performed for movement, I
think it would be nice to do some heavy refactoring of this. ...some day.