Mildly reorganize orc.des and crypt.des
Mostly some quick headers and splitting decorative vaults from
monster vaults. We're not very close to the minimum branch-unique
decor count for either branch (somewhat arbitrarily: 20 for one
floor, 27 for two, and 54 for four), but there's decorative vaults
inbound for both of these, and it'd make future assessment for
such be much faster.