Emphasize only intelligent monsters activate net traps
Whether or not this is a weird behaviour we entirely want to keep,
we might as well clarify the current specific behaviours of net
traps (always player-centric, only human-intelligence monsters
activte it) in their description.

I'm of two minds to keep this- it's extremely idiosyncratic and
unexpected compared to other traps, but removing it would make two
ossuaries and Zot:5 and Tomb all toss even more nets indefinitely
at players. Might be worth considering a limited activation count
and some context of only monsters activating it, in 0.32.