Fix monsters refusing to use pyre arrow / meph breath on other monsters
This is a very hacky stop-gap fix because I don't want to do the deeper
code refactoring that would be involved in a better fix after feature
freeze.

But the gist is that monsters would refuse to use these against other
monsters because the tracer says that they would inflict 0 damage (even
though the primary reason to use them is for their post-hit effect)

Currently, because AC is rolled during monster-vs-monster tracer checks,
low damage attacks can randomly be ruled irrelevant, even when their
side-effects would be very relevant. In practice, many of these do enough
damage that this is unlikely to occur (ie: quicksilver bolt, bolt of
devastation, and klown pies all do 3d20), but this may also be an issue
with some zaps with sufficiently low impact damage.

In future, it would be good to unify all 'apply a non-damaging effect from
a non-enchantment beam' into a single place that can check if a target is
eligable and then apply it, instead of the scattershot approach of checking
things in multiple different places (and sometimes via string comparison!)
This could even include the great mess that is monster cloud code!

But for now, this appears to fix the most immediate problem, and hoppefully
shouldn't cause any new ones.