Placeholder: rename + retile hard mode ice cave entries
This distinguishes easy and hard ice caves explicitly in game by making the
latter's entry feature be "glacial archways" to "a glacial chasm". This
also adds a new, higher-detail tile by Sastreii.

Currently, this is only a vault-defined feature name using a fragile hack
for the current branch-based difficulty split (as of fc7f089). In the next
version, we would like to make easy and hard mode portals (ice caves or
other future settings) lead to entirely different branches that happen to
share the same maps (e.g. said "Glacial Chasm"), rather than having the
current opaque hard cut-off. Adding new branches is beyond the scope of
currently being in feature freeze, however, so this is a somewhat
incomplete adjustment for now. It should still be an improvement, in that
it explains to newer players that there's any signifncat difficulty split
at all.