Nudge up kobold blastminer ranged damage (Monkooky)
While these are meant to be anti-melee monster experiments, the current low
damage of blastminer arbalests (less than orc warriors with arbalests)
makes it entirely plausible for sufficient AC characters to kill them via
poison darts rather than closing in and risking getting bombarded. This is
an unideal state for a monster as deep as Orc and involves some kiting,
so they're getting a slight nudge up of ranged damage (but not Bombard
damage) to try and discourage such tactics. Bombard is still a much bigger
focus of threat for the monster, anyway.