Catch a missing silver rune (CarefulOdds, Darby)
Might be a good idea for objstat to also track runes just in case anybody
does another big overhaul of all the V:$ quadrant subvaults or anything
else extremely long. Also, I feel obliged to say it's kind of dumb how
long we've kept the tradition of "O" being the rune glyph / branch entrance
glyph since it was not longer hardcoded to dynamically turn into the right
entrance / rune so very very long ago and it's so easy to read it as a 0.