Avoid a rare crash with eleionoma/witch wand creation (PleasingFungus)
The exact cause is still a bit obscure, but it seems like if they drop a
wand that is your last unID'd wand, and that wand was also on your shopping
list, it can cause a level excursion that wipes out the still-dying
monster's data before their death is finished processing.

I'm still not satisfied with this, as it's unclear why it only triggers for
the LAST wand (and not 100% of the time, even then), but this still seems
to greatly improve the situation.