Autopickup elemental evokers by default
Excluding the horn of Geryon, because it doesn't currently exist on
the \ menu, and we want players to be able to manually disable these
if desired.

Possibly the horn should be enabled in both places, since it is treated
internally as an xp evoker, but I'm not touching that here.