Fix a slew of frightening issues with pacified monster pathing
Pacified monsters were intended to path directly to the nearest level exit
and then leave the level. Sometimes they did this! And some other times,
deeply broken things happened. Where do I even begin...

First of all, there was a small chance each turn for a pacified monster to
give up on whatever exit it was moving towards (1-in-40 for human
intelligence monsters and 1-in-120 otherwise). The comments implied that
this was intended to help 'stuck' monsters look for a different staircase
instead, but the code never let them ACTUALLY look for a different one.
Instead, it would cause them to just start wandering vaguely in the general
vacinity of the nearest exit instead (possibly more than LoS range away
from it). Maybe they would randomly step on it at some point! But probably
not. (Also, if they were far enough from the exit at this point and
mindless, they would likely just wander around where they already were)

Also, if they rolled the random chance to stop moving towards an exit, that
exit would also become unusable to all subsequently pacified things (until
either ALL exits on a floor got marked that way, or the code hit a 1-in-20
chance when trying to find an exit for a pacified monster to regenerate
the cache from scratch). So they could give up, and then also cause future
monsters to ignore nearby stairs, too.

Also, for some reason, the code considered teleport traps as valid 'exits'
from the level, but it was only possible to trigger them 'leaving' by
stepping on one if they were also wearing -tele items (since otherwise they
would teleport before the 'leave level' check happened. If they landed
nearby, they would keep on walking into the trap over and over again,
trying desperately to accomplish something useful. If they DID have a -tele
item, on the other hand, stepping onto the trap would cause them to cease
to exist without even a message.

And most terrifyingly of all: the code used a cache of level exits on a
floor to avoid having to search for all of them each time a monster was
pacified. BUT THIS CACHE WAS NOT DISCARDED IF YOU LEFT THE FLOOR. This
meant that pacified monsters would try to path towards where exits were ON
A PREVIOUS FLOOR, and not the current one. And if they actually reached the
coordinates of one of those, they would simply vanish without a message.
As near as I can tell, this has likely been broken in this way for a very,
very long time.

Anyway! The code is different now. The cache is properly regenerated on
each new floor, monsters actually attempt to do a pathfinding check when
looking for an exit (to avoid the very real original problem of just being
stuck staring at an exit across water until they rolled the slim chance to
started wandering in place instead). They will even look for another exit
if the one they are moving towards becomes inaccessible during their
journey (detected somewhat crudely by 'has passed a turn without a valid
step forward', which isn't perfect, but is a hell of a lot improved).

If there are somehow NO exits they can path towards at all, they will
properly give up on trying and just wander around. The odds of being in
a completely disconnected spot (ie: on the other side of deep water) and
then this changing later (eg: by something teleporting them) feels slim
enough not to worry about it.

Also, teleport traps no longer count as valid exits. (Shafts and Gateways
to the Infinite Horrors of the Abyss still do, because they're funny)