Fix pacified good neutrals not acting like they were pacified
This is currently done to nearby orcs when you first accept a Beogh
conversion. I assume it was intended that they actually become pacified
or the could wouldn't have called mons_pacify(), but in practice this
didn't actually work, since for some reason monsters needed to have BOTH
the pacified flag AND be true neutral.