New Fugue and Blazeheart spell icons (CanOfWorms)
Both of them are good work, as usual.

Re-alphabetized a little of summoning and necromancy in dc-spells, while I
was adding new spells there. A lot of these tiles listings need some heavy
reorganizing, honestly.

Wereblood's icon is now used for monster Battlecry, as something rather
important to notice on a rather wide variety of different monsters, and
as something which reasonably fits as a shouting face.