Don't lie about Yred bound soul hp
It is perhaps surprising to some people that bound soul hp doesn't care
about the hp of the original monster, but (like all dervied undead) is
purely a multiplier of the original's HD.

8d68a812144 misleadingly made bound souls' description state the original's
hp, (which was much less than their actual hp for many souls, while being
also much MORE than various lategame unique souls).

Yred is slated for significant changes in 0.32, and more transparency for
how much hp a bound soul will have is part of that. But in the meantime,
xv should at least reflect the current state more accurately.