Tweak Sculpt Simulacrum's numbers
Reaction to this spell has been a bit mixed, with many believing that the
situational nature of the delay isn't worth the payoff. There's been
discussion of revamping a few aspects of the spell, but since we're in
feature freeze, I am going to just make modest numerical changes.

Originally the ice blocks were deliberately fragile, since I thought that
might give them interesting situational considerations, but it feels like
the delay alone provides that. I have now made them quite durable instead,
so that it's less likely your attempt to create summons will be distrupted
by damage - and in fact they may sometimes offer a defensive benefit.

I've also increased the summon duration from 2 to 3. It explicitly did not
want the summons to be so long-lived that the typical use-case was to
create them from the tail end of one battle so that you could bring them
into the next one. That is how necromany operates and I wanted this to be
as different-feeling as possible in that regard. But 3 is still pretty
common for regular summons, so hopefully fine.

We'll see how this incarnation feels across the length of 0.31