Catch lost Lair gems, adjust a Spider gem
baedc58 did fix placing too many runes in the Arachne spider end (since I
missed the corner placements in afdcd2b), but also easily places the gem
quite far away from the rune. With some thorough clean-up of the vault
header, I've also reconfigured the vault slightly to have a proper
arrangement of gems placements close to rune placements.

Also, the spreee of edits in 866d48a accidentally made wormcave fail to
place the gem on occasion, if the double-small-Lair-end workings chose it
as the second (but not the first) of those double-Lair-ends. Additionally,
if Crawl fails to place two smaller Lair ends as it rarely does, then it
wouldn't place the Lair gem at all. Both of these have been fixed, the
latter by placing it in the first Lair end the game places just in case
it doesn't succeed in placing the second.