Don't let spectral weapons get free hits every now and again
Actual triggered attacks by spectral weapons were handled by
spectral_weapon_fineff, but it was still possible for the spectral weapon
monster to randomly path into the player (or, presumably, something else).
And since all it takes in Crawl logic to initiate an attack is 'trying to
move into something we're not aligned with'), this would cause an attack.
Potentially multiple attacks in a row, in fact.

Since we never really want fight_melee to happen for spectral weapons, this
specifically blocks that.

(We could just reduce their speed to 0, but the sight of them dancing back
and forth feels a minor shame to lose, I think)