Make skeletons eternal
It seems weird that skeletons rot away in the presumably finite time
that the player spends within the dungeon, and it's cool to leave
more of a trace of your presence when retracing your tracks. Since
skeletons are purely decorative now, let's leave em around.

Some argument for removing the ridiculous and complex corpse decay
system entirely, but I don't have the heart.