fix: Serpent of Hell placement in Dis (Sar)
The subvault dis_grunt of iron_city_of_dis had bad substitutions that
could sometimes replace the glyph used for the SoH with some other
definition. I've reworked the substitutions to (1) always place the SoH
if Dis is its target branch, (2) randomize placement between the two
locations for the SoH the map uses, and (3) retain the slight
randomization of the spawn location in the rune chamber regardless of
whether SoH places on this map.