Implement monster foxfire (for illusions and ghosts)
Marshlight already used basically identical code to foxfire under the hood
(just with a key for cosmetic changes to it). To make this spell less
strong at the same HD, monsters cast it with half power compared to
marshlight, but it otherwise works the same way.

(Of course, this still scales up much more than the player spell at high
HD, but that's just how monster spells work in general)