Searing Ray monster implementation
Just for ghosts/illusions for the moment, but I think this one has real
potential on some other monsters (I have arcanists in mind in particular).
The dynamics of a channel that can be interrupted by the player moving
(either out of range or to block the beam with other monsters) feel like a
good one.

With this in mind, the monster version also lasts longer at higher power
than the player themselves can reach. If someone dislikes this, there's
always the possibility of a foxfire/marshlight split with a fancier name
for the arcanist version.