Let monsters cast Manifold Assault (and refactor a few things)
Beyond letting monsters use the spell in general, it refactors Autumn
Katana's code so that it isn't mostly-copying manifold assault code in a
separate way.

Additionally, this refactors the special handling that both of these (and
Wu Jian martial attacks) used for calculating weapon delay, so that it now
relies only player_speed() and not you.time_taken. Commits dating back
many years now suggest that it always should have worked this way, and I
have not seen any bad results in testing all of these things with some
rigor. Frankly, I'm of the opinon that you.time_taken should NEVER have any
meaning in the middle of any player action, and only after the player's
turn is complete.)