Catch a missing Depths gem (GnomeSayin)
evilmike_zot_entry_gauntlet doesn't use the same standardized function
of zot_entry_setup like the other zot entry vaults do because it uses
a different monster set than the others. As such, it was missed in
866d48a, which was mostly checking the presence of that function being
called. I edited Zot gem placement in 2f004bd, but I missed out on this
one also because I was just editing the use of that function. (I also
didn't catch it in afdcd2b adjusting a lot of gem placement, because it
wasn't one of the rune branch uses as I was worried about at the time).

Sorry, everybody who missed out on the gem this way. I'll fit another
review of smoky and midnight gems into the next big vault review for the
early start of 0.32.