Adjust monster regeneration enchantment and add Regenerate Other
ENCH_REGENERATION was only used as part of monster Trog's Hand (which was
itself used only by spriggan berserkers). Since it merely did the
equivalent of setting the M_FAST_REGEN flag, this made berserkers heal
exactly 1 hp per turn, instead of the ~0.48 they would have healed without
it. It did last a long time, but that is frankly a pathetic amount of
healing.

ENCH_REGENERATION has now been co-opted for another simple monster spell.
It now applies regeneration on TOP of whatever natural regen a monster has
(instead of replacing it) and heals 5% of the monster's max health + 3
each turn. This should let it scale a little to the target's own toughness
and be something the player can actually feel.

Regenerate Other is a simple use of this (intended for priest apostles at
the moment) that simply applies this buff to a random injured ally in
sight.

Since this also significantly buffs Trog's Hand, I have cut its duration to
less than half. (But I'm not convinced it is even that interesting or
relevant on spriggan berserkers, who are already nearly magic-immune in the
first place, and can't even use it if they go berserk first, which they
often will. Considering cutting it from them and maybe finding a new home
for it in future.)