feat(webserver): Add game_id to rcfile_contents
When webtiles clients request an rc file's contents, they send a
`get_rc` message with a `game_id` property. The webtiles server responds
with a `rcfiles_contents` message that has a `contents` property.

Becuase the response doesn't include `game_id`, the client has to treat
this as a stateful request.

This isn't a problem for the current webtiles client, which stores the
state in the UI. But it add complexity to the design of future clients.