Fix minor bugs with set_nearest_monster_foe and refactor slightly
The iterator for letting monsters pick other monsters as foes, for some
reason, used LOS_DEFAULT_RANGE, regardless of the player's current vision
range. This meant, for instance, that a monster at the extreme end of a
barachi's sight couldn't engage another monster from that distance.

The more concerning scenario - monsters still looking 7 tiles away for
targets if you're a kobold - turned out not to matter for another reason:
global los is cached using the player's LoS range, so calls to
monster::can_see() would still fail for overly far monsters - they would
just be called against targets that they would NEVER be able to see.

With that in mind, I have removed a redundent los check in _mons_check_foe,
corrected this bug, and renamed an argument for clarity.