Make Beogh/Yred recall instant again
With the number of permanent allies of Beogh capped at 3, incremental
recall felt rather silly. Sometimes it got all 3 of them in one turn and
sometimes it took 2 (with additional messages about starting/stopping
recall). This feels unnecessary.

Yred still can have a larger number of zombies at once, but given that
they are bound to a single floor and the mass armies that originally
prompted incremental recall are a thing of the past, I think we can safely
trim this code out and simplify.