Don't let allies repeatedly cast blink/battlesphere with nothing around
It was mildly amusing when a Yred bound soul might do this, but it should
be a lot more regular ocurrance for allies to have these spells with
Beogh now. And an ally blinking away from you right as you autotravel to
another level, so that they miss travelling through the stairs with you, is
more than a little awkward.