Add M_CAUTIOUS to make (selective) spellcasters less likely to approach
This is intended foremost for orc apostle wizards, but I think there are a
few other places where it may be appropriate (such as glowing orange
brains, whose spellset is especially hurt by waltzing up to the player
without even a melee attack)

The way it works is that if an M_CAUTIOUS monster doesn't opt to cast a
spell that turn (or use a wand, reaching attacking, etc.) but had a valid
spell it COULD have cast, and is not already in melee range of something,
it has a 2/3 chance to wait in place instead of moving forward. It's sort
of like having a cantrip that it won't ever use in place of attacking.

I am hoping this will make wizard apostles a bit better at not flinging
themselves into the jaws of death.