Allow followers to enter/be recall into portal vaults
The original veto for this appears to long predate interlevel recall and
I can't help but wonder if it was an attempt to prevent followers being
lost forever inside portals. In any case, that time has long passed and I
see no good reason to forbid a player from using apostles or bound souls
in portal levels.