Add Divine Armament, another spell to help round out apostle summons
Drawing upon Beogh's long history of directly granting weapons to their
followers, this spell summons a single dancing weapon. The dancing weapon
an individual apostle can summon will be identical between uses of this
spell can will be either an axe or a polearm (both of which I think are
the more interesting options for the player to have access to)

At low power, it summons either a plain war axe or trident. At medium
power, it summons either a battleaxe or glaive of electrocution. And at
max power, it summons either an executioner's axe or bardiche of speed.
Dancing weapon power also increases between these, so that the actual
power of the summon shows larger differences between tiers (and the bottom
tier is suitable to offer to the player fairly early)

Even weak dancing weapons could be dangerous early enemies for many
players, but I think Beogh worshippers' access to smite makes them uniquely
capable of dealing with these, and it will help to give priest apostles
more variety.