Fix dactions not being processed after level peek (Darby)
Each level stores an index into the global list of scheduled dactions
that reflects 'the last daction that was done on this level'. The idea
being that whenever you reenter a level where you've already been, it will
use this index to process only the 'new' dactions that have been added
since you were there last.

The problem is that (for some reason), the dactions_done wasn't actually
being marshalled properly, but the game was assuming that it was impossible
to save a floor without all dactions being caught up. And this is
distinctly NOT the case in at least one situation (and maybe more): level
excursions!

Essentially, making a level excursion to any floor would cause any dactions
scheduled (but not yet processed on that floor) to never happen there.
An easy way to demonstrate this is by pulling Kirke to another floor, away
from her hogs, and killing her there. Then peeking back at the old level.
If you did this, and then returned for real, the hogs left behind would
never transform back, since the excursion had made the game save that
floor's dactions_done as 'total number of dactions that exist'.

This potentially affects countless things, in a wide variety of stuations.
Even doing things with the shopping list triggers level excursions, which
would have invalidated daction state on the floor with the shop, I'm
pretty sure.

This fix is a very simple one, and seems to work (and feels like what this
code always should have been). But it's possible I'm overlooking something,
of course, and it can't be as simple as it seems. Fingers crossed?