Make groups of monsters handle diagonal chokepoints more consistently
Sometimes a monster that can attack the player will instead walk to the
side to let a friend move forward and help. For example, when fighting a
pack of yaks near the corner of a corridor

: #####
: #.YYY
: #@###
: #.###

the lead yak has a 50% chance of moving west into the corner instead of
attacking the player.

This behavior didn't work right in one particular type of situation:

: ###YY
: ###YY
: ..Y##
: .@.##
: ...##

It used to be that the lead yak would never move south or west in such
a situation, even though in the similar situation with an extra wall

: ###YY
: ###YY
: ..Y##
: .@###
: ...##

it would have the expected 50% chance of moving west. (There was a similar
issue as well with the player moved one square north in the preceding two
diagrams.)

This commit tries to fix this, so you will no longer be safe fighting one
yak at a time in the second diagram above. Other behavior is hopefully
unchanged.

Note that in the situation

: ###YY
: ###YY
: .@Y##
: .#.##
: ...##

the yak will still never move south - this is because currently monsters
will only change their move by a 45 degree turn, not a 90 degree turn as
would be needed here. The goal of this commit isn't to make groups of
monsters as smart as possible, but just to fix a single weirdness/bug with
the old behavior.