Don't let Throw Ally throw tentacles (Lici)
If an iron giant throws the tentacle tip, it will cause a disconnecting
sever. If it throws the tentacle segments, an unattached tentacle segment
will sit on the floor for one turn before disappearing the next. Either
way, this is definitely undesired behaviour we can avoid in the slim cases
of Zot trap eldritch tentacles or charm / scrolls of summoning plus Malign
Gateway in Dis and a tiny number of vaults.