remove tentacle submerge, and the rest of submerge
Submerge as a mechanic introduces very unique special cases that must be guarded against throughout the code. The untargettability it grants requires a special check in several targetters, and the ability to temporarily coexist with monsters makes removing the enchantment very complicated.

Right now the only monsters that can submerge are tentacle segments, for whom the ability is fairly uninteresting. Preserving submerge for tentacle segments is decidedly not worth the technical debt it incurs.

This does not touch the tile drawing code; it may still be possible to see 'submerged' tentacles if a tentacle somehow gets disconnected. There may be vestigial code for drawing other submerged entities.