Change how rCorr interacts with corrosion stacks on the player
Protecting yourself against corrosion feels bad because of how
inconsistent it is.

Each stack of corrosion on the players reduces AC and slay by 1.
Quadrupled most sources of corrosion to compensate. Having corrosion
resistance will halve the number of stacks applied to the player
instead of a coinflip chance to prevent the status entirely.