refactor: generalize some feature description code
2c8f65ca5f32ac introduced some code to try to allow redefined features
to be conditioned on layouts, for the sake of Dis statues. This code
made various assumptions like: there is always a layout, a level always
has vaults, and a layout is always at position 0 of the level vault
vector. bc18a39e1d331e tried to fix some of these assumptions, but
didn't really handle the case where a level has no vaults at all (can be
true in the abyss), which leads to an empty level vault vector.
(Accessing position 0 of such a vector may fail to crash under various
compiler and runtime circumstances...)

This commit refactors to (i) reuse previously existing, slightly more
robust code to find the vault at a position, and (ii) adds a function to
try to more robustly find a layout if one exists. It's unclear to me
whether the position 0 assumption was or wasn't safe if there is a
layout (it wasn't guaranteed, but may have been true in practice), but
this code doesn't make it. As far as I can tell this works with the dis
statue case.