Be more consistant about announcing Beogh conversion offers
The old code clearly intended the player be guaranteed to have a priest
speak to them (and print a message about how to convert) the first time
they were eligable. But since this was in behaviour_event() it wasn't
actually guaranteed that the priest would do this on any turn. In
particular, it seemed very unlikely on turns they cast spells (though they
would sometimes not bother for many turns in a row for unclear other
reasons).

I have pulled this outside of monster behavior code entirely. When the
player qualifies for the first time, they will reliably be informed at the
start of their turn. (Priests still have the old 1/10th chance to mention
conversion at any other eligable time)