Lower the contrast on bedeviled crystal tiles (PleasingFungus)
The fancy graphics on top of a wall tile is already plenty without needing
to quite catch the eye so much for being just a retiled / redescribed wall,
I suppose. The more colourful originals will sit in the git history anyway
for others' potential use, and the desaturating also sells corruption in
its own fashion.