Fix players being naked and refactor
The addition of a mirrored weapon tile category caused a bunch of code that
calculates what order to draw player paperdoll parts to work incorrectly.

And trying to fix it uncovered FOUR DIFFERENT PLACES that were duplicating
a handmade list and logic for determining this order. I have consolidated
these into a single function, and also attempted to reduce some other
duplicated logic in the process. (I think a little more couple be shaved
off, but it was less clear whether some of the other differences were
important or not)