Add some apostle names
All of them are references to mythical or historical defeated converts of
kings, warriors, retainers, or wild men, alluding a little to the new
defeat and recruit system attached to orcish apostles. Possibly a little
more could be done with this, possibly we'll just get other sorts of
references snuck in eventually.

Also, remove Harm as an apostle name, since it collides weirdly with the
actual brand we've had for a long while now.