Slow down sif piety decay
Halve Sif piety decay rate to make it the same as other gods with slow
piety decay.

Sif Muna is notorious for being the hardest god in the game to reach 6*
in piety with. Although Sif has some fun abilities, the player is often
discouraged from using them because of how limited Sif piety is
compared to other gods.