Make dismissing divine allies on penance a fineffect
If you read ?immo around orcish allies and something hit one, you would be
blamed for the resulting flame cloud and given penance *in the middle of
an attack*, resulting in deleting the attacked monster before the attack
was done, ultimately resulting in a crash. I suspect this is the same cause
as some TSO I've seen periodically, so I am moving all divine ally
dismissal into a fineffect. This technically results in the player being
able to get more penance at a time (by hitting multiple summons at once
before they disappear), but a solution to that might be somewhat complex
and I'm not sure it's necessary.

I also removed some old code to protect against a crash that should no
longer be necessary.