Interrupt travel when player allies are attacked from out of sight
Currently, while visible animations and messages happen when allies are
shot from out of player LoS, these do not interrupt travel. This can be
somewhat unpleasant for zombies and is deeply bad for apostles (whose
deaths is a major event). Now the player should be interrupted by default
when an ally is attacked by something the player cannot see.

This uses the ME_WHACK behaviour_event() reaction, which should cover a
great deal of offensive things used against them, but possibly not
everything.

I am a little concerned at this making it harder to use travel to flee a
battle where your zombies are holding the line, but the positive QoL from
this change is obvious. We'll see if any of this ends up being a problem.